+++ CHANGELOG.md
@@ -3,7 +3,16 @@
 ## New Components
 
 * Confirm
+* Grid
 * Animated Menu Button
+
+* Input Grid - Breaking change
+
+Input Grid now requires a onRowDelete prop and onRowUpdate prop instead of a deleteRowHandler prop and updateRowHandler prop
+
+## Minor
+
+* Input Grid is now decorated by common grid
 
 # 0.1.1